move login fields code out of main activity and into email view
