+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/Routes.kt
@@ -1,6 +1,8 @@
 package app.omnivore.omnivore
 
 sealed class Routes(val route: String) {
-    object Splash : Routes("Splash")
+    object Root : Routes("Root")
+    object Home : Routes("Home")
+    object Welcome : Routes("Welcome")
     object EmailLogin : Routes("EmailLogin")
 }
